requirements.txt
asyncio==3.4.3
dependency-injector==4.43.0
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
langchain==0.3.7
langchain-openai==0.2.8